adding params to route

+++ src/Utlis/ApiRequest.js
@@ -1,25 +1,32 @@
 const api = {
-  get: async (url) =>
-    fetch(url).then(async (data) => {
-      if (data.status === 200) return await data.json();
-      else return null;
-    }),
-  delete: async (url) =>
-    fetch(url, { method: "DELETE" }).then(async (data) => {
-      if (data.status === 200) return await data.json();
-      else return null;
-    }),
-  post: async (url, body) =>
-    fetch(url, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(body),
-    }).then(async (data) => {
-      if (data.status === 200) return await data.json();
-      else return null;
-    }),
+  get: async (url) => {
+    if (url)
+      return await fetch(url).then(async (data) => {
+        if (data.status === 200) return await data.json();
+        else return null;
+      });
+  },
+  delete: async (url) => {
+    if (url)
+      return await fetch(url, { method: "DELETE" }).then(async (data) => {
+        if (data.status === 200) return await data.json();
+        else return null;
+      });
+  },
+  post: async (url, body) => {
+    if (body)
+      return await fetch(url, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(body),
+      }).then(async (data) => {
+        if (data.status === 200) return await data.json();
+        else return null;
+      });
+    else return null;
+  },
 };
 
 export default api;
